Add character background test; separate Info files
- Complete test for character background.
- Separate Info config files to enable jest mocks of specific config
files. This allows testing all scenarios. For instance, character backgrounds can be mocked up to incorporate one background per test case instead of only using the default single background.
