+++ config/Info/Abilities.js
@@ -1,0 +1,3 @@
+export default [
+  'strength', 'dexterity', 'constitution', 'intelligence', 'wisdom', 'charisma',
+];
