+++ config/Info/Alignments.js
@@ -1,0 +1,4 @@
+export default [
+  'Lawful Good', 'Lawful Neutral', 'Lawful Evil', 'Neutral Good', 'True Neutral', 'Neutral Evil',
+  'Chaotic Good', 'Chaotic Neutral', 'Chaotic Evil',
+];
