+++ config/Info/BaseSkills.js
@@ -1,0 +1,110 @@
+export default {
+  acrobatics: {
+    skillLabel: 'Acrobatics',
+    modifier: 0,
+    ability: 'dexterity',
+    proficient: false,
+  },
+  animalHandling: {
+    skillLabel: 'Animal Handling',
+    modifier: 0,
+    ability: 'wisdom',
+    proficient: false,
+  },
+  arcana: {
+    skillLabel: 'Arcana',
+    modifier: 0,
+    ability: 'intelligence',
+    proficient: false,
+  },
+  athletics: {
+    skillLabel: 'Athletics',
+    modifier: 0,
+    ability: 'strength',
+    proficient: false,
+  },
+  deception: {
+    skillLabel: 'Deception',
+    modifier: 0,
+    ability: 'charisma',
+    proficient: false,
+  },
+  history: {
+    skillLabel: 'History',
+    modifier: 0,
+    ability: 'intelligence',
+    proficient: false,
+  },
+  insight: {
+    skillLabel: 'Insight',
+    modifier: 0,
+    ability: 'wisdom',
+    proficient: false,
+  },
+  intimidation: {
+    skillLabel: 'Intimidation',
+    modifier: 0,
+    ability: 'charisma',
+    proficient: false,
+  },
+  investigation: {
+    skillLabel: 'Investigation',
+    modifier: 0,
+    ability: 'intelligence',
+    proficient: false,
+  },
+  medicine: {
+    skillLabel: 'Medicine',
+    modifier: 0,
+    ability: 'wisdom',
+    proficient: false,
+  },
+  nature: {
+    skillLabel: 'Nature',
+    modifier: 0,
+    ability: 'intelligence',
+    proficient: false,
+  },
+  perception: {
+    skillLabel: 'Perception',
+    modifier: 0,
+    ability: 'wisdom',
+    proficient: false,
+  },
+  performance: {
+    skillLabel: 'Performance',
+    modifier: 0,
+    ability: 'charisma',
+    proficient: false,
+  },
+  persuasion: {
+    skillLabel: 'Persuasion',
+    modifier: 0,
+    ability: 'charisma',
+    proficient: false,
+  },
+  religion: {
+    skillLabel: 'Religion',
+    modifier: 0,
+    ability: 'intelligence',
+    proficient: false,
+  },
+  sleightOfHand: {
+    skillLabel: 'Sleight of Hand',
+    modifier: 0,
+    ability: 'dexterity',
+    proficient: false,
+  },
+  stealth: {
+    skillLabel: 'Stealth',
+    modifier: 0,
+    ability: 'dexterity',
+    proficient: false,
+  },
+  survival: {
+    skillLabel: 'Survival',
+    modifier: 0,
+    ability: 'wisdom',
+    proficient: false,
+  },
+};
