+++ config/Info/Languages.js
@@ -1,0 +1,22 @@
+export default {
+  standard: [
+    { language: 'common', speakers: ['humans'], script: 'common' },
+    { language: 'dwarvish', speakers: ['dwarves'], script: 'dwarvish' },
+    { language: 'elvish', speakers: ['elves'], script: 'elvish' },
+    { language: 'giant', speakers: ['ogres', 'giants'], script: 'dwarvish' },
+    { language: 'gnomish', speakers: ['gnomes'], script: 'dwarvish' },
+    { language: 'goblin', speakers: ['goblinoids'], script: 'dwarvish' },
+    { language: 'halfling', speakers: ['halflings'], script: 'common' },
+    { language: 'orc', speakers: ['orcs'], script: 'dwarvish' },
+  ],
+  exotic: [
+    { language: 'abyssal', speakers: ['demons'], script: 'common' },
+    { language: 'celestial', speakers: ['celestials'], script: 'dwarvish' },
+    { language: 'draconic', speakers: ['dragons', 'dragonborn'], script: 'elvish' },
+    { language: 'deep speech', speakers: ['mind flayers', 'beholders'], script: 'none' },
+    { language: 'infernal', speakers: ['devils'], script: 'infernal' },
+    { language: 'primordial', speakers: ['elementals'], script: 'dwarvish' },
+    { language: 'sylvan', speakers: ['fey creatures'], script: 'elvish' },
+    { language: 'undercommon', speakers: ['underdark traders'], script: 'elvish' },
+  ],
+};
